anbar_pro/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
psycopg[binary]==3.1.19
PyJWT==2.9.0
pandas==2.2.2
openpyxl==3.1.5
gunicorn==22.0.0
whitenoise==6.6.0
